+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {RouterModule, Routes} from "@angular/router";
-import {LoginComponent} from "./login/login.component";
+import {LoginComponent} from "./component/login/login.component";
 
 const routes: Routes = [
   //{ path: '', component: HomeComponent },
