+++ src/app/component/profile/profile.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { AuthenticationService, UserDto } from '../service/authentication.service';
+import { AuthenticationService, UserDto } from '../../service/authentication.service';
 import { Router } from '@angular/router';
 @Component({
   selector: 'app-profile',
