+++ src/app/component/quiz-list/quiz-list.component.css
