+++ src/app/component/scan/scan.component.html
