+++ src/app/component/scan/scan.component.spec.ts
